# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gym Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/setup:
    get:
      operationId: getSetupStatus
      tags: [auth]
      summary: Check if first-admin setup is needed
      responses:
        "200":
          description: Setup status
          content:
            application/json:
              schema:
                type: object
                properties:
                  needsSetup: { type: boolean }
                required: [needsSetup]
    post:
      operationId: runSetup
      tags: [auth]
      summary: Create the first admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                fullName: { type: string }
              required: [username, password, fullName]
      responses:
        "200":
          description: Token + user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
              required: [username, password]
      responses:
        "200":
          description: Token + user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user profile with role and permissions
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Not authenticated
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword: { type: string }
                newPassword: { type: string }
                confirmPassword: { type: string }
              required: [currentPassword, newPassword, confirmPassword]
      responses:
        "200":
          description: Password changed
...
components:
  schemas:
...
    PagePermissions:
      type: object
      description: Per-page and per-feature permissions for a user
      properties:
        dashboard: { type: boolean }
        members: { type: boolean }
        plans: { type: boolean }
        staff: { type: boolean }
        payroll: { type: boolean }
        payments: { type: boolean }
        vouchers: { type: boolean }
        accounts: { type: boolean }
        stock: { type: boolean }
        sales: { type: boolean }
        settings: { type: boolean }
        viewCost: { type: boolean }
        viewProfit: { type: boolean }
        viewAccounting: { type: boolean }
        manageStaff: { type: boolean }
        manageSettings: { type: boolean }
        managePayroll: { type: boolean }
        manageInventory: { type: boolean }
        manageMembers: { type: boolean }
        managePlans: { type: boolean }
      required:
        - dashboard
        - members
        - plans
        - staff
        - payroll
        - payments
        - vouchers
        - accounts
        - stock
        - sales
        - settings
        - viewCost
        - viewProfit
        - viewAccounting
        - manageStaff
        - manageSettings
        - managePayroll
        - manageInventory
        - manageMembers
        - managePlans
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/CurrentUser"
      required: [token, user]
...
    CurrentUser:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, staff]
        status: { type: string }
        permissions:
          $ref: "#/components/schemas/PagePermissions"
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - username
        - name
        - role
        - status
        - permissions
...